chore(tool): consolidate profile package tests into profile_test.go
Fourth package in the test-consolidation pass for #1128. profile_extra_test.go
had no matching source file; every function it tested (Start, Stop,
writeHeapProfile, mergeType, Merge) is declared in profile.go, whose
existing profile_test.go already held the base coverage for those
functions. Folded into profile_test.go.

This is a pure move: no test was added, removed, renamed, or edited.
The {name, pass/fail} set is identical before and after the merge.
